Convert #run to use splat opts
Allows for a more flexible run options structure, with the intention to
eventually support piping.

The one concession with this changes is the what used to be `command`,
so a string with all/some/none of the params directly built in, will now
include the entire built command when outputting via `run!`.  This is to
avoid parsing the opts from within `#run!` a second time, but does mean
there is a small breaking change for those expecting the `message` on
`CommandResultError` to be consistent.