Add Open3.pipe_capture3
This method is meant to simulate the arg structure of `Open3.capture3`,
but support the built in pipe support of `Open3`.

Most of the code here is borrowed from the current implementations of
`Open3.capture3` and `Open3.popen3`, and slightly modified to support
the `.pipeline_run` method and it's input/output.

Something of note:  An opinionated decision was made for how ENV was
handled for this.  Basically, the env as passed into `run` is set as a
global env for every cmd that is part of the pipe.  In `.pipeline_run`,
you are technically allowed to set this on a per command basis, but this
really doesn't fit the current argument model of AwesomeSpawn.